refactor: extract some create-astro test utils
