+++ packages/create-astro/test/utils.js
@@ -1,0 +1,32 @@
+import { execa } from 'execa'
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+export const testDir = dirname(__filename);
+export const timeout = 5000;
+
+const createAstroError = new Error(
+	'Timed out waiting for create-astro to respond with expected output.'
+);
+
+export function promiseWithTimeout(testFn) {
+	return new Promise((resolve, reject) => {
+		const timeoutEvent = setTimeout(() => {
+			reject(createAstroError);
+		}, timeout);
+		function resolver() {
+			clearTimeout(timeoutEvent);
+			resolve();
+		}
+		testFn(resolver);
+	});
+}
+
+export function setup(args = []) {
+	const { stdout, stdin } = execa('../create-astro.mjs', args, { cwd: testDir });
+	return {
+		stdin,
+		stdout,
+	};
+}
